move prebuild code into individual files as a `scripts` in scripts dir
